Produce all query timings for query regardless of state
Instead of delaying production of state timings until each state is complete,
produce final duration for completed states, partial duration for current state,
and zero duration for future states.

This simplifies the code, and always gives up to date data for the current
state. This does not have a large impact as these stats are use for informational
purposes and for final events (which would have final data).