Fix interruption handling in Driver
The issue was discovered when testing Presto on Spark.

Spark emits the `interrupt` signal to cancel the task.

If the interrupt signal is emitted when the Driver is about to finish (before
the `isFinished` method is called) the execution never finishes.

The `finish` method tries to ackquire the lock that alwayws returns `false`
when the interrupted flag is set. Since the `false` is interpreted as a failure
to aquire the lock, the finish is called again and again as it never succeeds.

Since the timeout for aquiring the lock is tiny (<=100ms) it is fine to try
to aquire the lock in a non interruptible manner to avoid taking care of non
trivial semantics of interrupting.