Collect and export setup and teardown query IDs in VerifierQueryEvent
The query IDs are collected once a query ID is received by the JDBC
client, not matter if the query succeeds or fails.

Previously, we're running teardown queries in a finally block after
VerifierQueryEvent is generated. We need to change that order in
order to export the IDs in the event.

Also, export setup query IDs and teardown query IDs for determinism
analysis, in determinism_analysis_details.